fix(agent-service): restructure SPI packages to <domain>.spi convention; fix catalog row count
Rename spi.executor -> executor.spi and spi.intercept -> intercept.spi to match
the established project convention where all SPI packages end in '.spi'.

Files updated:
- Java: package declarations + Javadoc FQN refs in 8 files (git mv'd from spi/executor/, spi/intercept/)
- agent-service/module-metadata.yaml: spi_packages entries
- docs/dfx/agent-service.yaml: spi_packages entries
- docs/contracts/contract-catalog.md: 5 FQN rows + count 40->45
- architecture/docs/L1/agent-service/spi-appendix.md: FQN table + parity check text
- architecture/docs/L1/agent-service/development.md: package tree listing
- architecture/docs/L1/agent-service/features/engine-dispatch-execution.md: FQN ref
- architecture/features/features.dsl: saa.devPaths in 2 FEAT- elements
- docs/adr/0155-agent-service-l1-v1-2-internal-module-design.yaml: decision block
