+++ agent-service/src/main/java/com/huawei/ascend/service/runtime/executor/spi/package-info.java
@@ -5,11 +5,11 @@
  * Status at this commit: design_only — interface declared, no production
  * implementation. Reference impls land in a follow-up impl-mode wave.</p>
  *
- * <p>The {@link com.huawei.ascend.service.runtime.spi.executor.ExecutorAdapter}
+ * <p>The {@link com.huawei.ascend.service.runtime.executor.spi.ExecutorAdapter}
  * SPI unifies three Agent forms — Native (in-process platform-bean DI),
  * Third-party (in-process framework-bridge replacement), and Remote
  * (out-of-process A2A protocol client) — behind a single execute contract.
- * {@link com.huawei.ascend.service.runtime.spi.executor.InjectionMode}
+ * {@link com.huawei.ascend.service.runtime.executor.spi.InjectionMode}
  * captures the wiring choice an adapter declares at registration time.</p>
  */
-package com.huawei.ascend.service.runtime.spi.executor;
+package com.huawei.ascend.service.runtime.executor.spi;
