+++ architecture/docs/L1/agent-service/features/engine-dispatch-execution.md
@@ -22,7 +22,7 @@ authority: "Absorbed from docs/logs/reviews/2026-05-26-agent-service-module-capa
 | AS-L1-F38 | Stream / partial-result handoff | AS-SC03, AS-SC09 | Hand token, step, tool progress, and partial result material to Access Layer / Internal Event Queue for projection instead of letting engines own client streams directly. | Inputs: engine stream chunk, progress event. Outputs: stream projection material, RunEvent material. | Internal Event Queue, Access Layer, Translation & Tool-Intercept. | Stream disconnect, partial-output schema mismatch, slow consumer. | OpenAI Agents stream events, LangGraph streaming. |
 | AS-L1-F39 | Adapter capability and version governance | AS-SC19, AS-SC20, AS-SC24 | Manage supported run modes, streaming, tool, checkpoint, S2C, delegation, resume schema, and version capabilities for engine / third-party adapters. | Inputs: adapter metadata, runtime probe, capability contract. Outputs: capability registry entry, version compatibility signal. | Access Layer, Task-Centric Control Layer, Session & Task Manager. | Stale adapter capability, resume schema drift, unsupported run mode. | A2A AgentCard, AgentScope metadata, LangChain4j capabilities. |
 
-| AS-L1-F60 | ExecutorAdapter SPI three forms | AS-SC15, AS-SC18 | Single SPI `ExecutorAdapter.execute(ExecutionRequest) : Flow.Publisher<AgentEvent>` carries three Agent shapes — Native (in-process platform-bean DI), Third-party (in-process framework-bridge replacement), Remote (out-of-process A2A protocol client). M4/upstream is form-agnostic. SPI: `com.huawei.ascend.service.runtime.spi.executor.ExecutorAdapter` (design_only). | Inputs: ExecutionRequest. Outputs: AgentEvent publisher. | Translation & Tool-Intercept (resource calls intercepted), Session & Task Manager (cursor source). | Adapter unavailable, vendor SDK leakage (rejected at registration time). | Temporal Activity, AgentScope `AgentHandler`, LangGraph4j `CompiledGraph.stream`. |
+| AS-L1-F60 | ExecutorAdapter SPI three forms | AS-SC15, AS-SC18 | Single SPI `ExecutorAdapter.execute(ExecutionRequest) : Flow.Publisher<AgentEvent>` carries three Agent shapes — Native (in-process platform-bean DI), Third-party (in-process framework-bridge replacement), Remote (out-of-process A2A protocol client). M4/upstream is form-agnostic. SPI: `com.huawei.ascend.service.runtime.executor.spi.ExecutorAdapter` (design_only). | Inputs: ExecutionRequest. Outputs: AgentEvent publisher. | Translation & Tool-Intercept (resource calls intercepted), Session & Task Manager (cursor source). | Adapter unavailable, vendor SDK leakage (rejected at registration time). | Temporal Activity, AgentScope `AgentHandler`, LangGraph4j `CompiledGraph.stream`. |
 | AS-L1-F61 | InjectionMode wiring declaration | AS-SC18 | Adapter declares `InjectionMode ∈ {NATIVE_DI, THIRD_PARTY_BRIDGE, EVENT_RELAY, NONE}` at registration; EDE-08 routes resource-call interception accordingly (sync direct, sync bridge, async event-relay, or no local interception for remote). | Inputs: adapter spec at startup. Outputs: registered wiring mode + per-Run resolution. | Translation & Tool-Intercept (consumes mode for routing). | Mode mismatch with adapter behaviour rejected at startup compliance scan. | Spring auto-configuration + bean-replacement patterns. |
 | AS-L1-F62 | Structured error class emission | AS-SC18, AS-SC11 | Adapter failures emit one of 14 `ErrorClass` values per `docs/contracts/error-class.v1.yaml`; no raw vendor exceptions cross the M5 boundary. M4 consumes the class to drive retry-vs-fail decisions. | Inputs: adapter exception. Outputs: structured `ErrorClass` value (one of MODEL_UNAVAILABLE / MODEL_RATE_LIMITED / INPUT_VALIDATION_ERROR / POLICY_DENIED / TOOL_FAILED / TOOL_TIMEOUT / MEMORY_FAILED / RETRIEVAL_FAILED / REMOTE_PROTOCOL_ERROR / CLIENT_CALLBACK_FAILED / INTERNAL_ERROR / CONTEXT_OVERFLOW / ADAPTER_UNAVAILABLE / DEADLINE_EXCEEDED). | Task-Centric Control Layer (retry policy consumer). | Class not enumerated (rejected at compile-time when enum lands); cross-class leakage detected by static scan. | Temporal Activity error taxonomy, Spring AI vendor error mapping. |
