+++ docs/dfx/agent-service.yaml
@@ -19,8 +19,8 @@ spi_packages:
   - com.huawei.ascend.service.session.spi               # rc23: ContextProjector SPI per ADR-0100
   - com.huawei.ascend.service.task.spi                  # rc23: TaskStateStore SPI per ADR-0100
   - com.huawei.ascend.service.agent.spi                 # rc43: Agent first-class entity SPI per ADR-0128 (HTTP-edge customer registration surface)
-  - com.huawei.ascend.service.runtime.spi.executor
-  - com.huawei.ascend.service.runtime.spi.intercept
+  - com.huawei.ascend.service.runtime.executor.spi
+  - com.huawei.ascend.service.runtime.intercept.spi
 
 releasability:
   artifact_coords: "io.github.springai-ascend:agent-service"
